Comments: restructure src by feature
