+++ projects/packages/comments/src/class-avatars.php
